refactor - use helper function instead of method in component
